+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - New #188: Add `canEditTag` field to `ChatMemberRestricted` and `ChatPermissions` types.
 - New #188: Add `SetChatMemberTag` method.
 - New #188: Add `canManageTags` field to `ChatMemberAdministrator` and `ChatAdministratorRights` types.
+- New #188: Add `canManageTags` parameter to `PromoteChatMember` method.
 
 ## 0.16 February 28, 2026